refactor(map): nest the map under a mapping module
Mechanical relocation ahead of the canonical-mutation work: `map/` moves
to `mapping/map/`, and `mapping/mod.rs` becomes the facade that will own
growth, culling and BA scheduling once those responsibilities move out of
the map itself.

`lib.rs` keeps `pub use mapping::map`, so every existing `crate::map::*`
import still resolves and no caller changes. All eleven relocated files
are byte-identical to their previous contents; the method and test name
sets are unchanged at 53 and 29.

Kept deliberately separate from the behaviour-changing tasks that follow.
