+++ crates/kornia-slam/src/mapping/map/keyframe.rs
