+++ crates/kornia-slam/src/mapping/map/map_point.rs
