+++ crates/kornia-slam/src/mapping/map/ops/bundle_adjustment.rs
