+++ crates/kornia-slam/src/mapping/map/ops/culling.rs
