+++ crates/kornia-slam/src/mapping/map/ops/selection.rs
